# src/public/api.yaml
openapi: 3.0.3
info:
  title: Swagger CookBook App - OpenAPI 3.0
  description: |-
    This is a CookBook App Server based on the OpenAPI 3.0 specification
  contact:
    email: nguyensytung2003@gmail.com
  version: 1.0.11
servers:
  - url: https://cook-book-app.vercel.app
...
paths:
...
  /post/comment/{idComment}:
    patch:
      tags:
        - posts  
      summary: Update comment
      description: Người dùng update nội dung comment
      operationId: updateContentComment
      security:
        - token: [ec6b8bb786643eff0ba4f230b279b7dc542f1b62d8e944d319c63467d3c8ed56]
      parameters:
        - name: idComment
          in: path
          description: ID của comment
          required: true
          schema:
            type: string
      requestBody:
        description: Message commnet you want update
        content:
          application/json:
            schema:
              properties:
                content:
                  type: string
                  example: This is content message update
          application/x-www-form-urlencoded:
            schema:
              properties:
                content:
                  type: string
                  example: This is content message update
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse200'
                  - type: object
                    properties:
                      data:
                        type: object
                        $ref: '#/components/schemas/Comment'
        '401':
          description: unauthorized or invalid token or token has expired
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse401'
        '404':
          description: data not found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse404'
        '500':
          description: message error operation or server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse500'
    delete:
      tags:
        - posts
      summary: Xóa comment 
      description: Xóa một comment bằng ID
      operationId: deleteComment
      security:
        - token: [ec6b8bb786643eff0ba4f230b279b7dc542f1b62d8e944d319c63467d3c8ed56]
      parameters:
        - name: idComment
          in: path
          description: ID của Comment
          required: true
          schema:
            type: string
      responses:
        '204':
          description: successful operation
        '401':
          description: unauthorized or invalid token or token has expired
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse401'
        '404':
          description: data not found
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse404'
        '500':
          description: message error operation or server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CommonResponse500'
# Phương 
...
components:
  schemas:
...
    Comment: 
      required:
        - postId
        - userId
        - content
        - parentCommentId
      type: object
      properties:
        _id:
          type: object
          example: 6524cff9765269140b3e5097
        postId: 
          type: string
          example: 6512935f878c6d59414639f1
        userId: 
          type: string
          example: 652383b89fb251f2338c36c9
        content:
          type: string
          example: Message comment
        parentCommentId:
          type: string
          example: null or _id of comment
        createdAt:
          type: string
          example: 2023-09-26T08:16:31.602+00:00
        updatedAt:
          type: string
          example: 2023-09-26T08:16:31.602+00:00
...
    CommonResponse200:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
          example: null
    CommonResponse401:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 401
        message:
          type: string
          example: unauthorized or invalid token or token has expired
        data:
          type: object
          example: null
    CommonResponse404:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 404
        message:
          type: string
          example: data not found
        data:
          type: object
          example: null
    CommonResponse500:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 500
        message:
          type: string
          example: message error operation or server error
        data:
          type: object
          example: null
  securitySchemes:
    token:
     type: apiKey
     name: Authorization
     in: header